Moved classes to the EvolutionTool.EvoDetermination namespace to improve orderly structure and shorten the names.
EvoCrtieriaMain and EvoCriteriaBonus have been created and implemented to create another layer. I chose this setup because it provides a common class for both main as bonus criteria while providing a read only main/bonus criteria object on the evolution criteria.
I think this setup will work really well if I create a reader or a property file setup for the evo criteria.

Next to that a compound tool method has been moved to the EvoToolbox to reduce the responsibilities of the Determinator.