The in-training to rookie evolution is now coverd by a regression test.
Fix a few bugs.

Renamed criteria to criterion when refering to a single criterion.

Added two objects to bundle the stats of the user digimon into main or bonus criteria. This allows the use of more specific input parameters.

Moved a couple of methods from the Determinator to the relevant toolboxes to create a more layered structure

Last but not leaste, became freaking happy about this amazing progress! Love it! :D